moved database.ts to a more logical location
